Noise Overhaul
Noise module used in minimal (https://github.com/caseman/noise) has pip install issues (https://github.com/caseman/noise/issues/32) in newer python versions and environments. As well, it seems largely unmaintained. Solution was to take pure python implementation listed in repo and implement it as standalone in minimal. If package were to be maintained or have issue resolved, it would be recommended to go back to that implementation.